Added method to reallocate surplus votes
Removed the SurplusVoteAllocator trait and just started with
a simple implementation. There are a bunch of ways of doing
this and I'm not sure how they all work and what the common
interface should be.

Added various methods for partitioning candidates into elected
and not elected, which needs to happen after any vote transfer.

The easiest way to detect if a surplus reallocation has elected
a new candidate is to count everything again, and then compare
the length of the elected candidate list.

The logic should be fairly similar after an elimination but
I've left this as is for now.